HP & Stamina Bar Implemented

+++ Wander/Assets/Scripts/UI/StaminaBar.cs
@@ -7,28 +7,14 @@ public class StaminaBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxStamina(int stamina)
+    public void SetMaxStamina(float stamina)
     {
         slider.maxValue = stamina;
         slider.value = stamina;
     }
 
-    public void SetStamina(int stamina)
+    public void SetStamina(float stamina)
     {
         slider.value = stamina;
     }
-
-    // public IEnumerator RegenStamina(int currentStamina, int maxStamina)
-    // {
-    //     int newStam = currentStamina;
-
-    //     yield return new WaitForSeconds(2);
-
-    //     while (currentStamina < maxStamina)
-    //     {
-    //         currentStamina += maxStamina / 400;
-    //         slider.value = currentStamina;
-    //         yield return regenTick;
-    //     }
-    // }
 }
